requirements.txt
openai>=1.13.3
qdrant-client>=1.7.0
python-dotenv>=1.0.0
schedule>=1.2.0
streamlit>=1.44.0
fastapi>=0.109.0
supabase>=2.15.0
python-jose>=3.4.0
pydantic>=2.7.4
uvicorn>=0.27.0
httpx>=0.27.0,<0.28.0
plotly>=6.0.0
pandas>=2.2.0
requests>=2.32.3
urllib3>=2.2.3
pydantic-core>=2.18.2 